requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
azure-core==1.29.4; python_version >= '3.7'
azure-cosmos==4.5.1
beautifulsoup4==4.12.2; python_version >= '3.6'
black==23.9.1
bs4==0.0.1
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
discord.py==2.3.2
exceptiongroup==1.1.3; python_version < '3.11'
frozenlist==1.4.0; python_full_version >= '3.8.0'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
openai==0.28.1
packaging==23.2; python_version >= '3.7'
pathspec==0.11.2; python_version >= '3.7'
platformdirs==3.11.0; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pytest-docker==2.0.1
pytest-mock==3.11.1
pytest==7.4.2
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_full_version >= '3.8.0'
tomli==2.0.1; python_version < '3.11'
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.8.0; python_full_version >= '3.8.0'
urllib3==2.0.6; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'